+++ lib/ASN1/Composite/RelativeDistinguishedName.php
@@ -11,16 +11,16 @@
 namespace FG\ASN1\Composite;
 
 use FG\ASN1\Exception\NotImplementedException;
-use FG\ASN1\Object;
+use FG\ASN1\ASNObject;
 use FG\ASN1\Universal\Set;
 
 class RelativeDistinguishedName extends Set
 {
     /**
      * @param string|\FG\ASN1\Universal\ObjectIdentifier $objIdentifierString
-     * @param \FG\ASN1\Object $value
+     * @param \FG\ASN1\ASNObject $value
      */
-    public function __construct($objIdentifierString, Object $value)
+    public function __construct($objIdentifierString, ASNObject $value)
     {
         // TODO: This does only support one element in the RelativeDistinguishedName Set but it it is defined as follows:
         // RelativeDistinguishedName ::= SET SIZE (1..MAX) OF AttributeTypeAndValue
@@ -29,7 +29,7 @@ class RelativeDistinguishedName extends Set
 
     public function getContent()
     {
-        /** @var \FG\ASN1\Object $firstObject */
+        /** @var \FG\ASN1\ASNObject $firstObject */
         $firstObject = $this->children[0];
         return $firstObject->__toString();
     }